constructed remaining request handlers
