+++ schema.psql
@@ -4,7 +4,7 @@ DROP TABLE IF EXISTS reviews, photos, characteristics, char_reviews CASCADE;
 
 -- Create Reviews Table
 CREATE TABLE reviews (
-  id INT PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   product_id INT,
   rating VARCHAR,
   review_temp_date BIGINT,
@@ -21,21 +21,21 @@ CREATE TABLE reviews (
 
 -- Create Photos Table
 CREATE TABLE photos (
-  photo_id INT PRIMARY KEY,
+  photo_id SERIAL PRIMARY KEY,
   review_id INT REFERENCES reviews(id),
   url TEXT
 );
 
 -- Create Characteristics Table
 CREATE TABLE characteristics (
-  characteristic_id INT PRIMARY KEY,
-  product_id INT REFERENCES reviews(product_id),
+  characteristic_id SERIAL PRIMARY KEY,
+  product_id INT,  -- Modify this line based on your data model
   name TEXT
 );
 
 -- Create Characteristic Reviews Table
 CREATE TABLE char_reviews (
-  id INT PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   characteristic_id INT REFERENCES characteristics(characteristic_id),
   review_id INT REFERENCES reviews(id),
   value INT
@@ -45,3 +45,4 @@ CREATE INDEX idx_characteristics_product_id ON characteristics(product_id);
 CREATE INDEX idx_char_reviews_characteristic_id ON char_reviews(characteristic_id);
 CREATE INDEX idx_photos_review_id ON photos(review_id);
 CREATE INDEX idx_char_reviews_review_id ON char_reviews(review_id);
+CREATE INDEX idx_reviews_product_id_review_date ON reviews(product_id, review_date, id);
